+++ internal/xsql/tx_fake.go
@@ -19,7 +19,7 @@ type txFake struct {
 
 func (tx *txFake) PrepareContext(ctx context.Context, query string) (_ driver.Stmt, finalErr error) {
 	onDone := trace.DatabaseSQLOnTxPrepare(tx.conn.trace, &ctx,
-		stack.FunctionID(""),
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/xsql.(*txFake).PrepareContext"),
 		&tx.beginCtx, tx, query,
 	)
 	defer func() {
@@ -58,7 +58,7 @@ func (tx *txFake) ID() string {
 
 func (tx *txFake) Commit() (err error) {
 	onDone := trace.DatabaseSQLOnTxCommit(tx.conn.trace, &tx.ctx,
-		stack.FunctionID(""),
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/xsql.(*txFake).Commit"),
 		tx,
 	)
 	defer func() {
@@ -76,7 +76,7 @@ func (tx *txFake) Commit() (err error) {
 
 func (tx *txFake) Rollback() (err error) {
 	onDone := trace.DatabaseSQLOnTxRollback(tx.conn.trace, &tx.ctx,
-		stack.FunctionID(""),
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/xsql.(*txFake).Rollback"),
 		tx,
 	)
 	defer func() {
@@ -97,7 +97,7 @@ func (tx *txFake) QueryContext(ctx context.Context, query string, args []driver.
 ) {
 	onDone := trace.DatabaseSQLOnTxQuery(
 		tx.conn.trace, &ctx,
-		stack.FunctionID(""),
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/xsql.(*txFake).QueryContext"),
 		tx.ctx, tx, query,
 	)
 	defer func() {
@@ -116,7 +116,7 @@ func (tx *txFake) ExecContext(ctx context.Context, query string, args []driver.N
 ) {
 	onDone := trace.DatabaseSQLOnTxExec(
 		tx.conn.trace, &ctx,
-		stack.FunctionID(""),
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/xsql.(*txFake).ExecContext"),
 		tx.ctx, tx, query,
 	)
 	defer func() {