+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added internal `gstack` codegen tool for filling `stack.FunctionID` with value from call stack
+
 ## v3.59.1
 * Fixed updating last usage timestamp for smart parking of the conns